moved the menu functions to its own library
